requirements.txt
vkbottle
tortoise-orm
python-dotenv
aiohttp